Cookies falling apart.
Crumbling, I suppose. Does an immigrating citizen need to know it's own
naturalization id? It should be enough for the leader to recognize that
the naturalized citizen's naturalization id has superseded an existing
id, that it is now newer than the existing government. This is how we
will know that a legislator has been restarted.

Ah, but how does a citizen know to reject any requests to participate in
a government? It needs a boolean to say that it is no longer a pod
person and has become a participant in the community.

See #113.