Work from log, not from queue.
Work from the top of the log instead of form the proposal queue. The
proposal queue is now just a queue, with proposals pushed and shifted.
The proposal queue is cleared when the legislator, as a leader, proposes
a new government, or if the legislator makes a new promise. It won't
have a proposal queue if it is not a leader. Thus, we no longer have a
problem with stale proposals.

Closes #147.
Closes #157.
Closes #160.