Route promise for sake of coverage.
Using the unrouted staging for promise messages creates an unreachable
branch in the `forwards` method. This is because there is nothing that
is ever placed in the envelopes attached to a specific route that is not
forwarded, there are never any lingering returns.

The pulse only ever needs to have a storage space for forwards, and with
the new logic for replication, it is possible that the collection of
envelopes specific is not necessary, except for use as an indicator that
the route is scheduled.

When the route is not specified for `promise` or `accepted` messages,
the envelopes are not used at all for returns.