Entirely reproducible.
Externalize time entirely. Replay a log of events from a legislator and
watch new events generated match the log exactly. With this we can put
this into production and test any failures.

Next steps are figuring out the meaningful, human readable events. When
something goes wrong, what are we going to look for? The current
government, the state of the current pulse, the most recent decisions
and decrees.

Also, you need to find a place where the log can be rotated because they
are going to be extremely chatty.