The `preferred` object is a model of incumbency.
We're trying to find members of the previous parliament to create a
quorum from the previous parliament. The new parliament must be composed
from a quorum of the incumbent parliament. We're not looking for
majority members specifically because if we have a quorum we have at
least one member of the majority, so we have the latest decisions.
Recall that if we have a decision, it may have been decreed, but the new
government will turn decisions into decrees. Finally, note that we are
always attempting to obtain a quorum, so if we have a parliament of
three members, our quorum will never shrink below two.