Linked-list based client.
The client should build a linked list. This allows the user to use any
entry as a memento. The client does not have to hold onto items in the
tree after they are made uniform. Iterating through the linked list
releases it's memory. This is a simpler data structure all around. The
user can hold onto state as long as they like and they do not have to
trigger a cleanup.

Closes #277.