Full `citizens` in message, `immigrated` property.
Add the immigrating member to `citizens` map in message so that event
handlers do not have to refer to the legislator and can just obtain the
necessary information from citizens. Citizens will be complete on
immigrate and on exile.

Add an `immigrated` property maintained by the `Legislator` to all of
the properties in `citizens`.