Delete exiled peers from leader.
We don't delete peer records from the leader so that the leader can
quickly exile anyone who's not responding, not waiting for the timeouts
to propagate from their representative.

We should, however, delete the peer record if the citizen has been
exiled. If we don't the citizen won't be able to immigrate again until
there is a leadership change.