Skip sync of rebooted citizen.
When a citizen reboots, it is possible that they will receive network
messages for the instance prior to reboot. In one particular race
condition, the representative of the rebooted citizen could be newly
assigned to represent that particular id, so it would have no existing
cached committed value for the id. It would synchronize to get the meta
information and see a zeroed committed value because the rebooted member
is uninitialized. It would then try to find an entry, but it would not
find the right one because the new member has yet to immigrate.