Delay always at least for a moment.
In order to simplify putting other citizens into the `_disappeared`
table we delay retry by at least one tick so that we don't have to
prevent retries using our pseudo-network. Retries will always wait on
an advance of the clock. (Although in practice that advance may occur
before we schedule the retry.)