More certain that well always be in range of log.
Now that I'm rejecting responses from any change in government or commit
strategy, I'm fairly confident that we'll never be asked to commit an
entry that has been shifted off the tail of the list. (Popped off the
tail?) If the promise of a commit is less than or equal to the promise
of the entry at the head of our list we will always compare the value of
the commit entry to our logged value assuming it will be there.