Do not update most committed on back sync.
On back sync we're going to get a government far more often than not.
That goverment may be one we proposed so we would be the leader. If that
government has as a majority member the fellow who's back syncing us
then we're going to try and send a fiat government to flush the
election. When we do we might find ourselves syncing with a peer who's
maximum committed value is not in our log. That triggers an assertion.
Thus, we don't update the maximum committed value on back sync.

Still more back sync issues to consider, though.