Remove ploymorphic Acceptor creation.
Not necessary. Don't actually work. Acceptor creation was only ever
invoked by a Recorder so we already know what sort of recording device
we have and what sort of recording device we need. It is always a
transition from Recorder to Acceptor. The Acceptor would never invoke
the `Paxos._prepare` function since it is already ready and willing to
handle a prepare message itself.