Replace Recorder rejections with assertions.
Add Docco to describe why certain messages are going to be filtered out
before reaching the Recorder. Remove the test to see if the register
entry's previous promise matches the Recorder's current entry's promise.
Add an assertion to ensure that the Recorder's current entry's promise
matches the promise of the top most entry in the log.