A queue of pings to let us know you're alive.
To terminate onboarding and start termination detection the chaperon
needs to know if the node functioning correctly. It can set a timeout
based on the new pinged queue, which isn't so much a queue as it is an
event you can listen to, since I expect the only application for a queue
listener would be to reset a timer which is a synchronous action.

We are pinged when we receive a message and our receiver is not in a
collapsed state. Note that minority members might continue synchronizing
with their constituents when they themselves are no longer reachable so
that the constituents won't know that things are dire until the minority
member has been reaped by the chaperon.

Also note that the leader synchronizes itself which is essentially a
noop but serves to ensure that the network connection is healthy. It
used to act as a keep alive when the network layer was a reverse proxy
sort of thing and it didn't go away which is nice because here we can
use it again to trigger our we-got-pinged message.