Set committed table for self when adding entry.
When we add an entry to the log we then set our most committed record
for ourselves. We used to let that get updated by the self synchronize,
but when a single member synod synchronize its constituents and itself
separately. When it synchronize constituents it will advance its minimum
propagated, but it will not include itself when determining the minimum.
If it manages to synchronize its constituents before it synchronizes
itself it will advance its minimum propagated promise to be greater than
the promise of its most committed record. Then when it synchronize
itself it not be able to build the synchronize message because the entry
most committed promise for itself because minimum propagated was
calculated and the entry was shifted out of the log for garbage
collection.