Mark Paxos messages as non-constituent.
That there where no false constituent flags on messages lead me to
believe that this was not in use at all, or that it was merely a way to
distinguish between constituent synchronization and keep-alive self
pings. It could be used to ignore a self-ping when the synod is a single
member synod, but it is really used to prevent minimum calculation when
the message is part of an election as a result of a collapsed
government. I'm marking constituent as false so that when I revisit this
again I'll be able to spot the distinction and hopefully recall it's
purpose and not remove it.