Unit test failed embarkation race.
An arriving islander provides a cookie so that when we sync with it the
first time we're certain that the instance we're initializing is the
same instance that initiated the arrival. If the islander where to crash
restart while we were adding it to the government, we'd use the cookie
to match know that the restarted islander is not the same one that
initiated the arrival. This would prevent the new islander for
initiating an arrival, rejected because we already have an entry. It
would also prevent syncing with the old arrival, so eventually the sync
would timeout and the islander would depart the island, we would remove
it from the list of constituents. Once removed an attempt to initiate
arrival by the islander would be accepted.